+++ crispy_forms_foundation/__init__.py
@@ -1,2 +1,2 @@
 """Django application to add 'django-crispy-forms' layout objects for 'Foundation for sites'"""  # noqa: E501
-__version__ = '0.6.0'
+__version__ = '0.6.1-pre.1'